all tests but on passed in CORE module (todo for christophe)

+++ src/testunits/module_BASE/HashTableTestSuite.h
@@ -574,7 +574,7 @@ namespace gum_tests {
       for ( unsigned int i = 0; i < size; ++i ) {
         float nb1 = i + i * 0.01f;
         float nb2 = float(i * 2);
-        t1.insert( std::pair<float, float>( nb1, nb2 ), i * 1000 + i );
+        t1.insert( std::pair<float, float>( nb1, nb2 ),i * 1000 + i);
       }
 
       std::vector<float> vect( size + 1 );
@@ -584,7 +584,7 @@ namespace gum_tests {
       }
 
       for ( gum::Idx i = 0; i < size; ++i ) {
-        std::pair<float, float> thepair( vect[i], 2.0 * i );
+        std::pair<float, float> thepair( vect[i], float(2.0 * i) );
         TS_ASSERT( t1.exists( thepair ) );
         TS_ASSERT_EQUALS( t1[thepair] , i * 1000 + i );
       }
@@ -628,7 +628,7 @@ namespace gum_tests {
       }
 
       for ( gum::Idx i = 0; i < size; ++i ) {
-        std::pair<double, double> thepair( vect[i], 2.0 * i );
+        std::pair<double, double> thepair( vect[i], double(2.0 * i ));
         TS_ASSERT( t1.exists( thepair ) );
         TS_ASSERT_EQUALS( t1[thepair] , i * 1000 + i );
       }
